doc(structure): move component docs into separate folder
