+++ docs/components/components/ComboBoxDocumentation.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {ComboBox, Option} from 'belle';
-import Code from './Code';
-import {propertyNameStyle, propertyDescriptionStyle} from './style';
+import Code from '../Code';
+import {propertyNameStyle, propertyDescriptionStyle} from '../../style';
 
 const animals = [
   {name: 'Abyssinian', description: 'The oldest breed of cat in the world!', image: 'images/abyssinian.jpg'},
