+++ docs/components/routes.js
@@ -2,24 +2,24 @@ import React from 'react';
 import { IndexRoute, Route } from 'react-router';
 
 import Base from './Base';
-import ComboBoxDocumentation from './ComboBoxDocumentation';
-import DatePickerDocumentation from './DatePickerDocumentation';
-import ButtonDocumentation from './ButtonDocumentation';
-import CardDocumentation from './CardDocumentation';
-import ChoiceDocumentation from './ChoiceDocumentation';
+import ComboBoxDocumentation from './components/ComboBoxDocumentation';
+import DatePickerDocumentation from './components/DatePickerDocumentation';
+import ButtonDocumentation from './components/ButtonDocumentation';
+import CardDocumentation from './components/CardDocumentation';
+import ChoiceDocumentation from './components/ChoiceDocumentation';
 import Configuration from './Configuration';
 import GettingStarted from './GettingStarted';
 import About from './About';
 import Home from './Home';
-import OptionDocumentation from './OptionDocumentation';
+import OptionDocumentation from './components/OptionDocumentation';
 import Philosophy from './Philosophy';
-import PlaceholderDocumentation from './PlaceholderDocumentation';
-import RatingDocumentation from './RatingDocumentation';
-import SelectDocumentation from './SelectDocumentation';
-import SeparatorDocumentation from './SeparatorDocumentation';
-import SpinnerDocumentation from './SpinnerDocumentation';
-import TextInputDocumentation from './TextInputDocumentation';
-import ToggleDocumentation from './ToggleDocumentation';
+import PlaceholderDocumentation from './components/PlaceholderDocumentation';
+import RatingDocumentation from './components/RatingDocumentation';
+import SelectDocumentation from './components/SelectDocumentation';
+import SeparatorDocumentation from './components/SeparatorDocumentation';
+import SpinnerDocumentation from './components/SpinnerDocumentation';
+import TextInputDocumentation from './components/TextInputDocumentation';
+import ToggleDocumentation from './components/ToggleDocumentation';
 import FormComponents from './FormComponents';
 import IntroducingBelleGuide from './guides/IntroducingBelle';
