+++ docs/style.js
@@ -4,14 +4,14 @@ const style = {
     padding: '0 20px 0 0',
     textAlign: 'left',
     verticalAlign: 'top',
-    color: 'grey'
+    color: 'grey',
   },
 
   propertyDescriptionStyle: {
     padding: 0,
     textAlign: 'left',
     verticalAlign: 'top',
-    paddingBottom: 30
+    paddingBottom: 30,
   }
 
 };
